fix: build

+++ src/format.cc
@@ -160,7 +160,7 @@ optional<Message> parse(BufferReader& buffer) {
             auto sub = Subscribe {
                 buffer.read<MessageFlags>().value(),
                 buffer.read<uint16_t>().value(),
-                0
+                uint16_t(0)
             };
             auto topic_type = static_cast<TopicIdType>(sub.flags.topic_id_type);
             if (topic_type == TopicIdType::Normal) {
